Automatically delete old auto-generated API reference pages when building documentation
By default, Sphinx provides warnings if an RST file is present in the `/docs/source` folder but the file is not included in the documentation. However, this sometimes results in warnings or errors if the name of a module is changed, since the Sphinx `autosummary` extension will automatically generate a documentation file for the new module name but leave the old file alone (thus resulting in a warning that the old file is not included in the documentation). To address this issue, commands were added to `conf.py` to delete the existing auto-generated documentation folder before building the documentation