Add options to configure how PyXX handles multiplying a `Unit` object by a constant
The meaning of multiplying or dividing a `Unit` object by a constant can be ambiguous. Therefore, to reduce confusion, this option was disabled by default, and a configuration option was added to allow users to manually enable it if desired. Detailed documentation was added about the configuration options. Note that the `ConstantUnitMathConventions` class was not added as a nested class within `Unit` for two reasons: (1) nested classes are not common in Python, so this adheres to general practices, and (2) this makes it easier to put the documentation for `ConstantUnitMathConventions` on its own page to improve readability of the project documentation